FirstCommit

+++ Hdn.Core.Architecture/Hdn.Core.Architecture.Api/Controllers/v1/TenantController.cs
@@ -1,10 +1,14 @@
-﻿using Hdn.Core.Architecture.Api.Controllers;
-using Hdn.Core.Architecture.Application.Dtos.Product;
+﻿using FluentValidation;
+using Hdn.Core.Architecture.Api.Controllers;
 using Hdn.Core.Architecture.Application.Dtos.Tenant;
+using Hdn.Core.Architecture.Application.Interfaces.Providers;
 using Hdn.Core.Architecture.Application.Interfaces.Services;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace hdn.net.architecture.WebApi.Controllers.v1
@@ -13,26 +17,81 @@ namespace hdn.net.architecture.WebApi.Controllers.v1
     public class TenantController : BaseApiController
     {
         private readonly ITenantService _tenantService;
+        private readonly IValidator<TenantRequest> _validator;
+        private readonly ILogger _logger;
+        private readonly IMessageProvider _messageProvider;
 
-        public TenantController(ITenantService tenantService)
+        public TenantController(
+            ITenantService tenantService,
+            IValidator<TenantRequest> validator,
+            ILogger logger,
+            IMessageProvider messageProvider)
         {
             _tenantService = tenantService;
-            //intanciar service aqui 
+            _validator = validator;
+            _logger = logger;
+            _messageProvider = messageProvider;
         }
 
         // GET: api/<controller>
         [HttpGet]
         public async Task<IActionResult> Get(int pageNumber, int pageSize)
         {
+            if (pageNumber == 0 || pageSize == 0)
+            {
+                var message = new List<string>();
 
-            return Ok(await _tenantService.Get(pageNumber, pageSize));
+                if (pageNumber == 0)
+                    message.Add(_messageProvider.RequiredParameter(nameof(pageNumber)));
+
+                if (pageSize == 0)
+                    message.Add(_messageProvider.RequiredParameter(nameof(pageSize)));
+
+                foreach (var m in message)
+                    _logger.LogWarning(m);
+
+                return new BadRequestObjectResult(message);
+            }
+
+            try
+            {
+                return Ok(await _tenantService.Get(pageNumber, pageSize));
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, $"[{ControllerContext.RouteData.Values["controller"]} - {ControllerContext.RouteData.Values["action"]}] - {e.Message}");
+                return new StatusCodeResult(StatusCodes.Status500InternalServerError);
+            }
+
+
         }
 
         // GET api/<controller>/5
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id)
         {
-            return Ok(await _tenantService.GetById(id));
+            if (id == 0)
+            {
+                var message = _messageProvider.RequiredParameter(nameof(id));
+
+                _logger.LogWarning(message);
+                return new BadRequestObjectResult(message);
+            }
+
+            try
+            {
+                var tenant = await _tenantService.GetById(id);
+
+                if (tenant is null)
+                    return new NotFoundObjectResult(_messageProvider.RegisterNotFound(nameof(id), id.ToString()));
+
+                return Ok(tenant);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, $"[{ControllerContext.RouteData.Values["controller"]} - {ControllerContext.RouteData.Values["action"]}] - {e.Message}");
+                return new StatusCodeResult(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // POST api/<controller>
@@ -40,15 +99,48 @@ namespace hdn.net.architecture.WebApi.Controllers.v1
         [Authorize]
         public async Task<IActionResult> Post([FromBody] TenantRequest tenant)
         {
-            return Ok(await _tenantService.Create(tenant));
+            var validationResult = _validator.Validate(tenant);
+
+            if (!validationResult.IsValid)
+            {
+                _logger.LogWarning(string.Join(string.Empty, validationResult.Errors));
+                return new BadRequestObjectResult(validationResult.Errors);
+            }
+
+            try
+            {
+                return Ok(await _tenantService.Create(tenant));
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, $"[{ControllerContext.RouteData.Values["controller"]} - {ControllerContext.RouteData.Values["action"]}] - {e.Message}");
+                return new StatusCodeResult(StatusCodes.Status500InternalServerError);
+            }
         }
 
         // PUT api/<controller>/
         [HttpPut]
         [Authorize]
         public async Task<IActionResult> Put([FromBody] TenantRequest tenant)
-        {           
-            return Ok(await _tenantService.Update(tenant));
+        {
+            var validationResult = _validator.Validate(tenant);
+
+            if (!validationResult.IsValid)
+            {
+                _logger.LogWarning(string.Join(string.Empty, validationResult.Errors));
+                return new BadRequestObjectResult(validationResult.Errors);
+            }
+
+            try
+            {
+                return Ok(await _tenantService.Update(tenant));
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, $"[{ControllerContext.RouteData.Values["controller"]} - {ControllerContext.RouteData.Values["action"]}] - {e.Message}");
+                return new StatusCodeResult(StatusCodes.Status500InternalServerError); ;
+            }
+
         }
 
         // DELETE api/<controller>/5
@@ -56,7 +148,23 @@ namespace hdn.net.architecture.WebApi.Controllers.v1
         [Authorize]
         public async Task<IActionResult> Delete(int id)
         {
-            return Ok(await _tenantService.Delete(id));
+            if (id == 0)
+            {
+                var message = _messageProvider.RequiredParameter(nameof(id));
+
+                _logger.LogWarning(message);
+                return new BadRequestObjectResult(message);
+            }
+
+            try
+            {
+                return Ok(await _tenantService.Delete(id));
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, $"[{ControllerContext.RouteData.Values["controller"]} - {ControllerContext.RouteData.Values["action"]}] - {e.Message}");
+                return new StatusCodeResult(StatusCodes.Status500InternalServerError);
+            }
         }
     }
 }
